+++ modules/DataBars.lua
@@ -24,13 +24,31 @@ function NoobTacoUIElv:SetupDatabars()
 
     -- DATATEXT ------------------------------------------------------------------
     E.db["datatexts"]["font"] = "Poppins-SemiBold"
-    E.db["datatexts"]["panels"]["RightMiniPanel"] = "Bags"
-    E.db["datatexts"]["panels"]["LeftCoordDtPanel"] = "Coords"
-    E.db["datatexts"]["panels"]["RightCoordDtPanel"] = "Broker_RunSpeed"
+    E.db["datatexts"]["fontSize"] = 13
+    E.db["datatexts"]["panels"]["Center"][1] = "Talent/Loot Specialization"
+    E.db["datatexts"]["panels"]["Center"][2] = ""
+    E.db["datatexts"]["panels"]["Center"][3] = "Time"
+    E.db["datatexts"]["panels"]["Center"][4] = ""
+    E.db["datatexts"]["panels"]["Center"][5] = "Volume"
+    E.db["datatexts"]["panels"]["Center"]["battleground"] = false
+    E.db["datatexts"]["panels"]["Center"]["enable"] = true
+    E.db["datatexts"]["panels"]["LeftChatDataPanel"][1] = "Guild"
+    E.db["datatexts"]["panels"]["LeftChatDataPanel"][2] = "Friends"
     E.db["datatexts"]["panels"]["LeftChatDataPanel"][3] = "Item Level"
-    E.db["datatexts"]["panels"]["LeftChatDataPanel"]["right"] = "Mana Regen"
-    E.db["datatexts"]["panels"]["LeftChatDataPanel"]["left"] = "Broker_RunSpeed"
-    E.db["datatexts"]["fontSize"] = 14
+    E.db["datatexts"]["panels"]["LeftChatDataPanel"]["backdrop"] = false
+    E.db["datatexts"]["panels"]["MinimapPanel"][1] = "Coords"
+    E.db["datatexts"]["panels"]["MinimapPanel"][2] = "MovementSpeed"
+    E.db["datatexts"]["panels"]["RightChatDataPanel"][1] = "Experience"
+    E.db["datatexts"]["panels"]["RightChatDataPanel"][2] = "System"
+    E.db["datatexts"]["panels"]["RightChatDataPanel"][3] = "Gold"
+    E.db["datatexts"]["panels"]["RightChatDataPanel"]["backdrop"] = false
+
+    E.global["datatexts"]["customPanels"]["Center"]["backdrop"] = false
+    E.global["datatexts"]["customPanels"]["Center"]["name"] = "Center"
+    E.global["datatexts"]["customPanels"]["Center"]["numPoints"] = 5
+    E.global["datatexts"]["customPanels"]["Center"]["width"] = 945
+    E.global["datatexts"]["settings"]["MovementSpeed"]["Label"] = "Speed"
+    E.global["datatexts"]["settings"]["MovementSpeed"]["decimalLength"] = 0
 
     -- DATABARS ------------------------------------------
     E.db["databars"]["threat"]["enable"] = false
@@ -40,20 +58,21 @@ function NoobTacoUIElv:SetupDatabars()
 
     E.db["databars"]["experience"]["textFormat"] = "NONE"
     E.db["databars"]["experience"]["font"] = "Poppins-Medium"
-    E.db["databars"]["experience"]["height"] = 222
-    E.db["databars"]["experience"]["hideAtMaxLevel"] = false
+    E.db["databars"]["experience"]["height"] = 10
+    E.db["databars"]["experience"]["hideAtMaxLevel"] = true
     E.db["databars"]["experience"]["hideInCombat"] = false
-    E.db["databars"]["experience"]["orientation"] = "VERTICAL"
+    E.db["databars"]["experience"]["orientation"] = "HORIZONTAL"
     E.db["databars"]["experience"]["questCompletedOnly"] = true
     E.db["databars"]["experience"]["showBubbles"] = false
-    E.db["databars"]["experience"]["width"] = 8
+    E.db["databars"]["experience"]["width"] = 385
+
     E.db["databars"]["reputation"]["enable"] = true
     E.db["databars"]["reputation"]["font"] = "Poppins-Medium"
-    E.db["databars"]["reputation"]["height"] = 222
+    E.db["databars"]["reputation"]["height"] = 8
     E.db["databars"]["reputation"]["hideInCombat"] = false
-    E.db["databars"]["reputation"]["orientation"] = "VERTICAL"
+    E.db["databars"]["reputation"]["orientation"] = "HORIZONTAL"
     E.db["databars"]["reputation"]["showBubbles"] = false
-    E.db["databars"]["reputation"]["width"] = 8
+    E.db["databars"]["reputation"]["width"] = 200
 
     E.db["databars"]["colors"]["experience"]["a"] = 1
     E.db["databars"]["colors"]["experience"]["r"] = 0.14901960784314